+++ Code/GraphMol/FileParsers/CDXMLParser.cpp
@@ -661,7 +661,9 @@ void visit_children(
           mols.pop_back();
           continue;
         }
-        MolOps::assignStereochemistry(*res, true, true, true);
+        bool cleanIt = true, force = true;
+        MolOps::assignStereochemistry(*res, cleanIt, force,
+                                      params.flagPossible);
       } else {
         MolOps::detectBondStereochemistry(*res);
       }