+++ Code/GraphMol/FileParsers/Mol2FileParser.cpp
@@ -951,10 +951,11 @@ std::unique_ptr<RWMol> MolFromMol2DataStream(std::istream &inStream,
     MolOps::cleanUp(*res);
 
     try {
-      // when we sanitize for mol2, we skip the cleanup organometallic step since it's
-      // not really compatible with the semantics of mol2 files
-      constexpr auto sanitizeFlags = MolOps::SanitizeFlags::SANITIZE_ALL ^
-                            MolOps::SanitizeFlags::SANITIZE_CLEANUP_ORGANOMETALLICS;
+      // when we sanitize for mol2, we skip the cleanup organometallic step
+      // since it's not really compatible with the semantics of mol2 files
+      constexpr auto sanitizeFlags =
+          MolOps::SanitizeFlags::SANITIZE_ALL ^
+          MolOps::SanitizeFlags::SANITIZE_CLEANUP_ORGANOMETALLICS;
       if (params.removeHs) {
         // Bond stereo detection must happen before H removal, or
         // else we might be removing stereogenic H atoms in double
@@ -965,7 +966,8 @@ std::unique_ptr<RWMol> MolFromMol2DataStream(std::istream &inStream,
         // rings in bond stereo detection, and another in
         // sanitization's SSSR symmetrization).
         unsigned int failedOp = 0;
-        MolOps::sanitizeMol(*res, failedOp, MolOps::SanitizeFlags::SANITIZE_CLEANUP);
+        MolOps::sanitizeMol(*res, failedOp,
+                            MolOps::SanitizeFlags::SANITIZE_CLEANUP);
         MolOps::detectBondStereochemistry(*res);
         MolOps::RemoveHsParameters rhp;
         bool sanitize = false;
@@ -986,7 +988,8 @@ std::unique_ptr<RWMol> MolFromMol2DataStream(std::istream &inStream,
     }
 
     res->updatePropertyCache(false);
-    MolOps::assignStereochemistry(*res, true, true);
+    bool cleanIt = true, force = true;
+    MolOps::assignStereochemistry(*res, cleanIt, force, params.flagPossible);
   }
 
   return res;